Fixed branching
When modifying the program counter by successfully executing a branch, we set the reference program counter we're adding to as the program counter of the JUST EXECUTED branch instruction. We only do this in the pipeline. The NO_PIPE section is adjusted so this change is only relevant in the pipeline section. JR instructions are unaffected; only BEQ and BZ are relevant here.

CURRENTLY: ALL THREE FUNCTIONAL MODES HAVE THE SAME INSTRUCTION STATS AT THE END.